Bug 1822619: fixed NullPointerException
